Переносит логику при клике на город в CityList
